requirements.txt
altgraph==0.17.4
anyio==4.8.0
atlassian-python-api==3.41.19
beautifulsoup4==4.12.3
blessed==1.20.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Deprecated==1.2.18
editor==1.6.6
gitdb==4.0.12
GitPython==3.1.37
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
iniconfig==2.1.0
inquirerpy==0.3.4
jmespath==1.0.1
oauthlib==3.2.2
ordered-set==4.1.0
packaging==24.2
pfzy==0.3.4
pluggy==1.5.0
progress==1.6
prompt_toolkit==3.0.50
pycparser==2.22
PyGithub==2.6.1
pyinstaller-hooks-contrib==2025.0
PyJWT==2.10.1
PyNaCl==1.5.0
pytest==8.3.5
python-dotenv==1.0.1
python-gitlab==5.3.1
PyYAML==6.0.2
readchar==4.2.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
runs==1.2.2
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
wrapt==1.17.2
xmod==1.8.1
zstandard==0.23.0
